Initial release

+++ src/edit.php
@@ -3,8 +3,10 @@ require 'db.php';
 require 'session.php';
 include 'finediff.php';
 
-$TITLE_CHANGE_PERMISSION = 1;
-$ARTICLE_DELETE_PERMISSION = 1;
+function removeScriptTags($text) {
+  return preg_replace('#<script(.*?)>(.*?)</script>#is', '', $text);
+}
+
 
 $article_title = empty($_GET['t']) ? $_POST['article-title'] : $_GET['t'];
 
@@ -44,19 +46,27 @@ $article_id = $row["id"];
 $article_content = $row["content"];
 $article_tags = $row["tags"];
 $article_hits = $row["hits"];
+$article_permission = intval($row["permission"]);
 
 $result->free();
 
+if($article_permission > $user_permission) {
+  $error = true;
+  $error_message = "이 지식을 편집하기 위한 권한이 부족합니다.";
+}
+
 // 저장하기
-if (isset($_POST['article-content'])) {
+if (isset($_POST['article-content']) && $article_permission <= $user_permission) {
 
   $article_content_previous = $article_content;
   $article_tags_previous = $article_tags;
 
-  $article_new_title = $_POST['article-new-title'];
-  $article_content = $_POST['article-content'];
-  $article_tags = $_POST['article-tags'];
+  $article_new_title = removeScriptTags($_POST['article-new-title']);
+  $article_content = removeScriptTags($_POST['article-content']);
+  $article_tags = removeScriptTags($_POST['article-tags']);
   $article_delete = isset($_POST['article-delete']);
+  $article_change_permission = isset($_POST['article-permission']);
+  $article_permission = intval($_POST['article-permission']);
 
   if (empty($article_title) || empty($article_id)) {
     exit();
@@ -69,21 +79,33 @@ if (isset($_POST['article-content'])) {
     if ($db->multi_query($sqlQuery) === TRUE) {
       header('Location: 404.php');
     }
-
     exit();
   }
 
   // 변경된 내용이 없으면 그냥 리다이렉트
   if (strlen($article_content) - strlen($article_content_previous) == 0) {
-    if (strcmp($article_content, $article_content_previous) == 0 && strcmp($article_tags, $article_tags_previous) == 0 && strcmp($article_title, $article_new_title) == 0) {
+    if (strcmp($article_content, $article_content_previous) == 0 
+      && strcmp($article_tags, $article_tags_previous) == 0 
+      && strcmp($article_title, $article_new_title) == 0 && !$article_change_permission) {
       header('Location: read.php?t='.$article_title."&update=2");
       exit();
     }
   }
 
+  // 자기보다 권한을 높게 설정할 수는 없음.
+  if ($article_permission > $user_permission) {
+    header('Location: read.php?t='.$article_title."&update=2");
+    exit();
+  }
+
   $sqlQuery = "UPDATE `$db_articles_table` SET ";
   $sqlQuery .= "`content`='".$db->real_escape_string($article_content)."', ";
   $sqlQuery .= "`tags`='".$db->real_escape_string($article_tags)."' ";
+
+  if ($article_change_permission) {
+    $sqlQuery .= ",`permission`='".$article_permission."' ";
+  }
+
   if (strlen($article_new_title) > 1  && $user_permission >= $TITLE_CHANGE_PERMISSION) {
     $titleChanged = true;
     $sqlQuery .= ",`title`='".$db->real_escape_string($article_new_title)."' ";
@@ -105,6 +127,9 @@ if (isset($_POST['article-content'])) {
     $sqlQuery .= (strlen($article_content) - strlen($article_content_previous)).");";
 
     if ($db->query($sqlQuery) === TRUE) {
+
+      store_log($db, $loggedin ? $user_name : $user_ip, "항목 수정", $sqlQuery);
+
       if ($titleChanged) {
         header('Location: read.php?t='.$article_new_title."&update=1");
       } else {
@@ -122,12 +147,12 @@ if (isset($_POST['article-content'])) {
 }
 $db->close();
 
-$title = $article_title." - 편집하기";
+$page_title = $article_title." - 편집하기";
+$page_location = "edit.php?t=".$article_title;
 include 'header.php';
 ?>
 
 <link href="libs/bootstrap-markdown/css/bootstrap-markdown.min.css" rel="stylesheet">
-
 <script src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
 <script src="libs/bootstrap-markdown/js/markdown.min.js"></script>
 <script src="libs/bootstrap-markdown/js/bootstrap-markdown.js"></script>
@@ -146,21 +171,20 @@ include 'header.php';
   <form action="edit.php" method="post">
 
     <?php
-    if ($user_permission >= $TITLE_CHANGE_PERMISSION) { 
+    if ($user_permission >= $TITLE_CHANGE_PERMISSION && $user_permission >= $article_permission) { 
       echo '<div class="collapse" id="edit-title">';
       echo '<input type="text" class="form-control input-lg" name="article-new-title" placeholder="이 지식의 제목" value="'.$article_title.'" aria-describedby="helpBlock">';
       echo '<span id="helpBlock" class="help-block"><em>지식 제목의 잦은 변경은 다른 사용자들에게 혼란을 줄 수 있습니다.</em></span></div>';   
     }
+    echo "<hr>";
+    if ($error) {
+      echo "<div class=\"alert alert-danger\" role=\"alert\">".$error_message."</div>";
+    }
     ?>
-    <hr>
     <div class="well well-sm">
       자세한 편집 방법은 <a href="read.php?t=편집 방법">편집 방법</a> 문서를 참조하세요.<br/>
       <em>지식에 대해 고의적인 훼손을 가하거나 악의적인 내용을 작성할 경우 차단될 수 있습니다.</em>
     </div>
-    <?php
-    if ($error) {
-      echo "<div class=\"alert alert-danger\" role=\"alert\">".$error_message."</div>";
-    }?>
   
     <div class="form-group">
       <label for="content">내용</label>
@@ -171,12 +195,35 @@ include 'header.php';
       <input type="text" name="article-tags" class="form-control"  id="tags" value="<?php echo $article_tags;?>">
     </div>
     <input type="hidden" name="article-title" value="<?php echo $article_title;?>">
+      <?php
+
+  if($user_permission >= $article_permission && $user_permission > 0) {
+        echo '
+      <div class="form-group">
+      <label for="id">편집 가능한 권한</label>    
+      <select name="article-permission" class="form-control" id="permission">
+        <option value="0" '.($article_permission == 0 ? "selected" : "").'>독자</option>
+        <option value="1" '.($article_permission == 1 ? "selected" : "").'>편집자</option>';
+
+        if($user_permission > 1) {
+          echo '<option value="2" '.($article_permission == 2 ? "selected" : "").'>중재자</option>';
+        }
+        if($user_permission > 2) {
+          echo '<option value="3" '.($article_permission == 3 ? "selected" : "").'>관리자</option>';
+        }
+      echo '</select></div>';
+}
+
+?>
+
+
     <div class="text-center">
       <?php
-      if ($user_permission >= $ARTICLE_DELETE_PERMISSION) { 
+      if ($user_permission >= $ARTICLE_DELETE_PERMISSION && $user_permission >= $article_permission) { 
+ 
         echo '<div class="checkbox"><label><input type="checkbox" name="article-delete" value="1" onclick="return deleteAlert(this);">이 지식 삭제하기</label></div>';
       }?>
-      <button type="submit" class="btn btn-default" onclick="window.onbeforeunload=null;">업데이트</button>
+      <button type="submit" class="btn btn-default <?php echo ($article_permission > $user_permission ) ? "disabled" : "";?>" onclick="window.onbeforeunload=null;">업데이트</button>
       <a href="read.php?t=<?php echo $article_title;?>" class="btn btn-danger" role="button">취소</a>
     </div>
   </form>
